step05 예외처리하는 기능 구현

+++ Spring05_Final/src/main/java/com/gura/spring05/file/service/FileServiceImpl.java
@@ -2,6 +2,7 @@ package com.gura.spring05.file.service;
 
 import java.io.File;
 import java.util.List;
+import java.util.concurrent.CancellationException;
 
 import javax.servlet.ServletContext;
 import javax.servlet.http.HttpServletRequest;
@@ -11,6 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.gura.spring05.exception.CanNotDeleteException;
 import com.gura.spring05.file.dao.FileDao;
 import com.gura.spring05.file.dto.FileDto;
 
@@ -128,6 +130,13 @@ public class FileServiceImpl implements FileService{
 		//2. 삭제할 파일의 정보를 읽어와서 삭제할 파일의 저장된 파일명을 얻어낸다.
 		FileDto dto=dao.getData(num);
 		String saveFileName=dto.getSaveFileName();
+		//파일 작성자와 로그인된 아이디가 다르면 예외를 발생시킨다.
+		String id=(String)request.getSession().getAttribute("id");
+		if(!id.equals(dto.getWriter())) {
+			//예외를 발생시켜 메소드가 정상 수행 되지 않도록 막기
+			throw new CanNotDeleteException();
+		}
+		
 		//3. DB 에서 파일 정보 삭제
 		dao.delete(num);
 		//4. 파일 시스템에서 파일 삭제
